refactor: replace relative imports with @ alias
