+++ src/__tests__/pages/About.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
 
-import { About } from '../pages/About.tsx';
+import { About } from '../../pages/About.tsx';
 
 describe('About', () => {
   it('renders the about page correctly', () => {
