+++ src/__tests__/state/context/ThemeProvider.test.tsx
@@ -1,7 +1,11 @@
 import { describe, it, expect, beforeEach } from 'vitest';
 
-import { createMockStore, renderWithProviders } from './utils/test-utils.tsx';
-import { ThemeProvider } from '../state/context/ThemeProvider.tsx';
+import { ThemeProvider } from '@/state/context/ThemeProvider.tsx';
+
+import {
+  createMockStore,
+  renderWithProviders,
+} from '../../utils/test-utils.tsx';
 
 describe('ThemeProvider', () => {
   beforeEach(() => {
