+++ src/pages/Home.tsx
@@ -1,12 +1,12 @@
 import { type FC, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
-import { ErrorBoundary } from '../features/error-boundary/ErrorBoundary.tsx';
-import { DetailsView } from '../svalka/DetailsView.tsx';
-import { FallbackUI } from '../svalka/FallbackUI.tsx';
-import { SearchResults } from '../svalka/SearchResults.tsx';
+import { ErrorBoundary } from '@/features/error-boundary/ErrorBoundary.tsx';
+import { DetailsView } from '@/svalka/DetailsView.tsx';
+import { FallbackUI } from '@/svalka/FallbackUI.tsx';
+import { SearchResults } from '@/svalka/SearchResults.tsx';
 
-import type { CharacterData } from '../utils/hooks/useCharacterData.ts';
+import type { CharacterData } from '@/utils/hooks/useCharacterData.ts';
 
 interface HomeProps {
   characterData: CharacterData;
